+++ backend/app/schemas/event_record_detail.py
@@ -7,14 +7,14 @@ from pydantic import BaseModel
 class EventRecordDetailBase(BaseModel):
     """Base schema for event record detail."""
 
-    heart_rate_min: Decimal | None = None
-    heart_rate_max: Decimal | None = None
+    heart_rate_min: int | None = None
+    heart_rate_max: int | None = None
     heart_rate_avg: Decimal | None = None
 
-    steps_min: Decimal | None = None
-    steps_max: Decimal | None = None
+    steps_min: int | None = None
+    steps_max: int | None = None
     steps_avg: Decimal | None = None
-    steps_total: Decimal | None = None
+    steps_total: int | None = None
 
     max_speed: Decimal | None = None
     max_watts: Decimal | None = None
@@ -22,19 +22,19 @@ class EventRecordDetailBase(BaseModel):
     average_speed: Decimal | None = None
     average_watts: Decimal | None = None
 
-    moving_time_seconds: Decimal | None = None
+    moving_time_seconds: int | None = None
     total_elevation_gain: Decimal | None = None
 
     elev_high: Decimal | None = None
     elev_low: Decimal | None = None
 
-    sleep_total_duration_minutes: Decimal | None = None
-    sleep_time_in_bed_minutes: Decimal | None = None
+    sleep_total_duration_minutes: int | None = None
+    sleep_time_in_bed_minutes: int | None = None
     sleep_efficiency_score: Decimal | None = None
-    sleep_deep_minutes: Decimal | None = None
-    sleep_rem_minutes: Decimal | None = None
-    sleep_light_minutes: Decimal | None = None
-    sleep_awake_minutes: Decimal | None = None
+    sleep_deep_minutes: int | None = None
+    sleep_rem_minutes: int | None = None
+    sleep_light_minutes: int | None = None
+    sleep_awake_minutes: int | None = None
 
 
 class EventRecordDetailCreate(EventRecordDetailBase):
